Ansi parser (#136)
* initial work to add ansi formatting to agent tasks view. upgrade dependencies

* make ansi changes to plugins

* changelog

* rename agent task table

* fix a couple bugs

* fix component name
